Learnable 4D positional embeddings
I call this the "axe murderer" update. The goal of this learning project was to get familiar with model engineering by custom-designing the architecture of a transformer to solve a task it is not architecturally aligned with through its implicit pattern-making (2D spacial reasoning). This required custom solutions and patches to work around any of the architectural contradictions, like order not being structure or sparse rewards. This has led to a patchwork of adaptations where excessive complexity folds in on itself. It invents an entire symbolic ontology of spatial agency through relation-modulated identity embeddings. The script has become too convoluted for meaningful analysis. The model performs well at everything except what it was supposed to do: navigating around walls.